+++ app/helpers/is-equal.js
@@ -1,0 +1,1 @@
+export { default, isEqual } from 'ember-bootstrap/helpers/is-equal';