+++ app/helpers/is-not.js
@@ -1,0 +1,1 @@
+export { default, isNot } from 'ember-bootstrap/helpers/is-not';