+++ index.js
@@ -41,12 +41,10 @@ module.exports = {
             app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.svg'), { destDir: '/fonts' });
             app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.ttf'), { destDir: '/fonts' });
             app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.woff'), { destDir: '/fonts' });
+            app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.woff2'), { destDir: '/fonts' });
         }
 
-
-        // Import transition support
         app.import(path.join(app.bowerDirectory, 'bootstrap/js/transition.js'));
-        //app.import(path.join(app.bowerDirectory, 'bootstrap/js/modal.js'));
 
     }
 };