+++ tests/dummy/app/router.js
@@ -10,9 +10,11 @@ Router.map(function() {
     this.route('button');
     this.route('forms');
     this.route('alert');
+    this.route('accordion');
+    this.route('collapse');
     this.route('modal', function() {
-        this.route('from-action');
-        this.route('from-route');
+      this.route('from-action');
+      this.route('from-route');
     });
 });