+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+<a name"0.4.0"></a>
+## 0.4.0 (2015-09-22)
+
+* **Collapse:**
+  * added bs-collapse component for collapsible content ([794b52d5](https://github.com/kaliber5/ember-bootstrap/commit/794b52d5f6ad1c53e6a468590d69c269d33c1da5))
+  * refactored bs-accordion-item to use bs-collapse internally ([58b0d868](https://github.com/kaliber5/ember-bootstrap/commit/58b0d868c58c2c5f29f6dcc99e1eccb357280b9e))
+
+<a name"0.3.0"></a>
+## 0.3.0 (2015-09-09)
+
+
+#### Features
+
+* **Accordion:**
+  * added accordion components ([f2bbaca8](https://github.com/kaliber5/ember-bootstrap/commit/f2bbaca8))
+
+<a name"0.2.0"></a>
+## 0.2.0 (2015-08-14)
+
+
+#### Features
+
+* **Ember:** Updated components to be compatible with Ember 1.13 (thanks to [jesenko](https://github.com/jesenko)) and 2.0. *NOTE: Ember 1.13 is the minimum required version for ember-bootstrap 0.2.0 and above*
+* **Tests:** Refactored all component tests to use the new integration test style
+* **Select:** added Ember.Select replacement component ([19144085](https://github.com/kaliber5/ember-bootstrap/commit/19144085))
+
+<a name"0.1.4"></a>
+### 0.1.4 (2015-08-05)
+
+
+#### Bug Fixes
+
+* **Components.FormElement:** Wrong classes for horizontal layout forms (14) ([1caba612](https://github.com/kaliber5/ember-bootstrap/commit/1caba612))
+
+
+<a name"0.1.3"></a>
+### 0.1.3 (2015-07-07)
+
+
+#### Bug Fixes
+
+* **Ember:** Updated to Ember-cli 0.2.7, fixed deprecation warnings in Ember 1.13 ([1fc78c4](https://github.com/kaliber5/ember-bootstrap/commit/1fc78c41cfd4f7f72095825e14c813d62d772dae)) 
+* **Ember:** Updated to Ember-cli 0.2.5 / Ember.js 1.12. Fixed test failures because of missi ([45e59f03](https://github.com/kaliber5/ember-bootstrap/commit/45e59f03))
+
+
+#### Features
+
+* **Form:** submitOnEnter submits the form when user presses enter, even if no submit button ([fcc9898b](https://github.com/kaliber5/ember-bootstrap/commit/fcc9898b))
+
+
 <a name"0.1.2"></a>
 ### 0.1.2 (2015-04-22)